requirements.txt
boto3==1.16.50
botocore==1.19.50
jmespath==0.10.0
python-dateutil==2.8.1
python-environ==0.4.54
s3transfer==0.3.3
six==1.15.0
urllib3==1.26.2
